Started working on introducing fallback
Began introducing the machinery necessary for falling back to a lower order, more diffusive, Riemann Solver when the change in density would lead to a negative value. Development of this feature is incomplete (it was put on hold after realizing that the feature was not necessary for the current science case).

Currently, the changes introduced in this commit recompute the fluxes on the entire local block during the partial timestep when negative densities are encountered. This would introduce inconsistencies in the fluxes on the outer edges of the outer zones between blocks. To make this consistent, we would need to have the entire domain fallback whenever negative densities are encountered.

In the subsequent commit, I worked on introducing fallback only for the interfaces of just the cells that have the negative densities. Development of this strategy was relatively far along and employed iteration masks. However, I have since realized that under this approach, we would need 2 more ghost zones (one for the predictor step and one for the full time-step) to prevent inconsistencies from developing between the edges of neighboring grids. Alternatively, you could communicate with the 14 neighbors (6 faces and 8 vertices) about whether fallback is required in the ghost zone.