Refactored MethodFrameTransform to rely on the improved scheduling of Method objects.
Since the origin_offset can no longer be updated every cycle, we had to stop tracking the origin_offset as a block property and instead track the last_updated_origin_offset and the last time when the the frame was updated. Instead of recording the time at the start of the cycle (and using it to compute the origin offset), the time at the end of the cycle is used. This choice was made because the reference frame update occurs in the last method call of a cycle (all other preceeding methods were applied using the preceeding velocity). Another benefit of this choice is that if an output is written at the end of this cycle, both the time of the last velocity update AND the last updated origin offset will reflect the actual simulation time AND the actual origin offset.