Made EnzoEFltArrayMap, which is a map that holds EFlt3DArray
EnzoEFltArrayMap is designed to help transition away from using Grouping objects within the VLCT integrator.

Also decided to commit to the pointer-like semantics of CelloArray. A const CelloArray always points to the same array. However the values are mutable. In the future, there should be some testing of CelloArray holding a const type.