Largely refactored EnzoRiemann to make use of maps instead of groupings and fields.
This is the first step in overhauling EnzoMethodMHDVlct to transition from Groupings to Maps. This overhaul will affords a number of benefits. These include:
1. Being more simple (the grouping system is already conceptually similar to a mapping, but it has an extra level of indirection)
2. It will decouple the algorithm from the underlying Cello machinery. This will make the code easier to understand (and I think it will make Flux Corrections easier to add). This will also make future ports to GPUs far simpler.
3. It will eventually allow us to reduce the overhead of allocating and deallocating memory for temporary fields for every block for every cycle (we could introduce a memory pool that reuses the same memory on each node for these operations during every cycle)

The Riemann Solver machinery now works with mappings instead of groups. Note that more refactoring will be necessary to refactor the general interface later on, so that it's better optimized for this strategy. Additionally, for performance purposes, we may move away from using maps altogether within the riemann solver (I currently envision using the maps primarily for source terms/eos/stuff where performance is not usually a bottleneck)