Introduced copy_to instance method to CelloArray.
This new method is introduced so that the elementwise assignment method can be removed from the codebase. This substantially simplifies the implementation of CelloArray. With this change, we no longer support a direct method for assigning a scalar value to an entire array. That's not a big loss, there were only 2 places in the entire codebase that used this feature (both were in EnzoInitialShockTube). These 2 places were updated to reflect the removal of this feature.

The website documentation for CelloArray and the CelloArray unit tests were updated to reflect this change